requirements.txt
httpx[http2]>=0.27.0
aiohttp==3.9.5
requests==2.32.3
urllib3==2.2.2
python-dotenv==1.0.1
matplotlib==3.9.0
numpy==1.26.4
psutil>=5.9.0
psycopg2-binary>=2.9.0
asyncpg>=0.27.0
uplink>=0.9.7
python-telegram-bot>=20.7
pyTelegramBotAPI>=4.14.0
setuptools>=40.0.0
# Add psutil if you plan to implement detailed CPU/memory tracking per library
# psutil==5.9.8 